Refactor and fix room ordering
* AbstractRoomOrdering and OrderByTag are now in their own file pairs; they also now derive from QObject
* doSetOrder() correctly reconfigures connections to m_roomOrder (actually closes #524).
