+++ client/models/roomlistmodel.h
@@ -19,120 +19,13 @@
 
 #pragma once
 
+#include "abstractroomordering.h"
 #include "../quaternionroom.h"
 #include <connection.h>
 #include <util.h>
 
 #include <QtCore/QAbstractItemModel>
 #include <QtCore/QMultiHash>
-
-struct RoomGroup
-{
-    QVariant key;
-    QVector<QMatrixClient::Room*> rooms;
-
-    bool operator==(const RoomGroup& other) const
-    {
-        return key == other.key;
-    }
-    bool operator!=(const RoomGroup& other) const
-    {
-        return !(*this == other);
-    }
-    bool operator==(const QVariant& otherCaption) const
-    {
-        return key == otherCaption;
-    }
-    bool operator!=(const QVariant& otherCaption) const
-    {
-        return !(*this == otherCaption);
-    }
-    friend bool operator==(const QVariant& otherCaption,
-                           const RoomGroup& group)
-    {
-        return group == otherCaption;
-    }
-    friend bool operator!=(const QVariant& otherCaption,
-                           const RoomGroup& group)
-    {
-        return !(group == otherCaption);
-    }
-};
-using RoomGroups = QVector<RoomGroup>;
-
-class RoomListModel;
-
-class AbstractRoomOrdering
-{
-    public:
-        using Room = QMatrixClient::Room;
-        using Connection = QMatrixClient::Connection;
-        using groups_t = QVariantList;
-
-        AbstractRoomOrdering(RoomListModel* m) : _model(m) { }
-        virtual ~AbstractRoomOrdering() = default;
-
-    public: // Overridables
-        virtual QString orderingName() const = 0;
-        virtual QVariant groupLabel(const RoomGroup& g) const = 0;
-        virtual bool groupLessThan(const RoomGroup& g1,
-                                   const QVariant& g2key) const = 0;
-        virtual bool roomLessThan(const QVariant& group,
-                                  const Room* r1, const Room* r2) const = 0;
-
-        virtual groups_t roomGroups(const Room* room) const = 0;
-        virtual void connectSignals(Connection* connection) = 0;
-        virtual void connectSignals(Room* room) = 0;
-
-    public:
-        using groupLessThan_closure_t =
-                std::function<bool(const RoomGroup&, const QVariant&)>;
-        groupLessThan_closure_t groupLessThanFactory() const;
-
-        using roomLessThan_closure_t =
-                std::function<bool(const Room*, const Room*)>;
-        roomLessThan_closure_t roomLessThanFactory(const QVariant& group) const;
-
-    protected:
-        const RoomListModel* model() const { return _model; }
-        RoomListModel* model() { return _model; }
-//        const RoomGroups& modelGroups() const;
-//        RoomGroups& modelGroups();
-
-        void updateGroups(Room* room);
-
-    private:
-        RoomListModel* _model;
-};
-
-class OrderByTag : public AbstractRoomOrdering
-{
-    public:
-        explicit OrderByTag(RoomListModel* m)
-            : AbstractRoomOrdering(m), tagsOrder(initTagsOrder())
-        { }
-
-        QString orderingName() const override { return QStringLiteral("tag"); }
-        QVariant groupLabel(const RoomGroup& g) const override;
-        bool groupLessThan(const RoomGroup& g1,
-                           const QVariant& g2key) const override;
-        bool roomLessThan(const QVariant& groupKey,
-                          const Room* r1, const Room* r2) const override;
-
-        groups_t roomGroups(const Room* room) const override;
-        void connectSignals(Connection* connection) override;
-        void connectSignals(Room* room) override;
-
-    private:
-        QStringList tagsOrder;
-
-        static const QString Invite;
-        static const QString DirectChat;
-        static const QString Untagged;
-        static const QString Left;
-
-        static QStringList initTagsOrder();
-};
 
 class RoomListModel: public QAbstractItemModel
 {
@@ -193,7 +86,7 @@ class RoomListModel: public QAbstractItemModel
 
         std::vector<ConnectionsGuard<QMatrixClient::Connection>> m_connections;
         RoomGroups m_roomGroups;
-        std::unique_ptr<AbstractRoomOrdering> m_roomOrder;
+        AbstractRoomOrdering* m_roomOrder = nullptr;
 
         QMultiHash<const Room*, QPersistentModelIndex> m_roomIndices;
